+++ src/elements/bomb.cpp
@@ -31,7 +31,7 @@ int update_BOMB(UPDATE_FUNC_ARGS) {
 						pmap[y][x] = 0;
 						for (nxj=-(rad+1); nxj<=(rad+1); nxj++)
 							for (nxi=-(rad+1); nxi<=(rad+1); nxi++)
-								if ((pow(nxi,2))/(pow((rad+1),2))+(pow(nxj,2))/(pow((rad+1),2))<=1) {
+								if ((pow((float)nxi,2))/(pow((float)(rad+1),2))+(pow((float)nxj,2))/(pow((float)(rad+1),2))<=1) {
 									nb = sim->create_part(-1, x+nxi, y+nxj, PT_BOMB);
 									if (nb!=-1) {
 										parts[nb].tmp = 1;
@@ -43,7 +43,7 @@ int update_BOMB(UPDATE_FUNC_ARGS) {
 								}
 						for (nxj=-rad; nxj<=rad; nxj++)
 							for (nxi=-rad; nxi<=rad; nxi++)
-								if ((pow(nxi,2))/(pow(rad,2))+(pow(nxj,2))/(pow(rad,2))<=1)
+								if ((pow((float)nxi,2))/(pow((float)rad,2))+(pow((float)nxj,2))/(pow((float)rad,2))<=1)
 									if ((pmap[y+nxj][x+nxi]&0xFF)!=PT_DMND && (pmap[y+nxj][x+nxi]&0xFF)!=PT_CLNE && (pmap[y+nxj][x+nxi]&0xFF)!=PT_PCLN && (pmap[y+nxj][x+nxi]&0xFF)!=PT_BCLN) {
 										sim->delete_part(x+nxi, y+nxj, 0);//it SHOULD kill anything but the exceptions above, doesn't seem to always work
 										sim->pv[(y+nxj)/CELL][(x+nxi)/CELL] += 0.1f;
